FlaskServer/requirements.txt
flask
flask_cors
python-dotenv
pillow
google-generativeai
google-api-python-client